requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-Limiter==3.7.0
Flask-Caching==2.3.0
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.1
bleach==6.1.0
Werkzeug==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
eventlet==0.36.1
